Conflicts:
	extra/scripts/zentyal-unit-tests
	main/core/ChangeLog
	main/monitor/ChangeLog
	main/monitor/debian/control
	main/remoteservices/ChangeLog
	main/remoteservices/src/EBox/RemoteServices.pm
	main/remoteservices/src/EBox/Reporter/Base.pm
	main/remoteservices/stubs/proxy-redirections.conf.mas

+++ main/openvpn/src/EBox/OpenVPN/Model/UploadClientBundle.pm
@@ -19,7 +19,7 @@ package EBox::OpenVPN::Model::UploadClientBundle;
 
 use base 'EBox::Model::DataForm::Action';
 
-use Error qw(:try);
+use TryCatch::Lite;
 
 use EBox::Global;
 use EBox::Gettext;
@@ -107,12 +107,11 @@ sub formSubmitted
     my $openvpn = EBox::Global->modInstance('openvpn');
     try {
         $openvpn->setClientConfFromBundle($clientName, $bundle);
+    } catch ($e) {
+        unlink $bundle if (-f $bundle);
+        $e->throw();
     }
-    finally {
-        if (-f $bundle) {
-            unlink $bundle;
-        }
-    };
+    unlink $bundle if (-f $bundle);
 }
 
 # Method: pageTitle
